requirements.txt
openai==1.12.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6 